cud books

+++ src/services/book.service.ts
@@ -3,11 +3,14 @@ import { Op } from 'sequelize';
 
 export class BookService {
     async addBook(bookData: BookCreationAttributes, user_role: string) {
+        if (!user_role || user_role === 'customer') {
+            throw new Error('Unauthorized: Only admin can modify books');
+        }
         return Book.create(bookData);
     }
 
     async findBook(id: string, user_role: string) {
-        let book: BookAttributes | null = null;
+        let book: BookAttributes | null = {} as BookAttributes;
         console.log('User role in findBook:', user_role);
         if (!user_role || user_role === 'customer') {
             book = await Book.findOne({
@@ -46,6 +49,9 @@ export class BookService {
     }
 
     async modifyBook(id: string, updatedData: Partial<BookAttributes>, user_role: string) {
+        if (!user_role || user_role === 'customer') {
+            throw new Error('Unauthorized: Only admin can modify books');
+        }
         const book = await Book.findByPk(id);
         if (!book) {
             return null;
@@ -54,6 +60,9 @@ export class BookService {
     }
 
     async removeBook(id: string, user_role: string) {
+        if (!user_role || user_role === 'customer') {
+            throw new Error('Unauthorized: Only admin can modify books');
+        }
         const deletedCount = await Book.destroy({ where: { id } });
         return deletedCount > 0;
     }
